Add an alternate method for creating IOP explanation objects.
This method separates the process for explaining a type from the
instantiation of explanations for concrete objects.
For example, you can define how to explain a Velocity object, and then
reuse that definition for Velocity objects located at different paths.
This method also uses relative paths, so there is no combinatorial
explosion when explaining objects that reside in sets.