Fixed bug where type names that had certain characters in them wouldn't be handled correctly.
I think the issue is that type names that included non-word chars weren't working. What a non-word char is appears to be platform dependent; on linux it's anything that isn't a number, letter, or underscore. So when I had a type name with a "*" in it, it didn't work. The fix is to wrap the constructed type names with quotes or braces so these other characters would be included.

(cherry picked from commit 096fd0cf4b9e6ee05f60be726d1f2aec761ad69e)